Add require for docker group users to ensure correct state ordering
This change tries to prevent the case where users are added to the docker group before said users are created as part of the same salt run. This should yield a better state application stability when a single run should yield the correct result (i.e. no need for reruns to resolve ordering problems)