Dnsmasq should strictly follow the nameserver order in resolv.conf
To prevent dnsmasq from choosing an upstream server, follow the order in resolv.conf instead.

dnsmasq docs:
By default, dnsmasq will send queries to any of the upstream servers it knows about and tries to
favour servers that are known to be up. Setting this flag forces dnsmasq to try each query
with each server strictly in the order they appear in /etc/resolv.conf